requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
isort==4.3.9
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.2.2
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
